IMS-VT: Modification of data usage aidl
Change data usage aidl interface to take parameter type long
instead of int

CRs-Fixed: 747694

+++ telecomm/java/android/telecom/InCallService.java
@@ -362,7 +362,7 @@ public abstract class InCallService extends Service {
              *
              * @param dataUsage The updated data usage.
              */
-            public abstract void onCallDataUsageChanged(int dataUsage);
+            public abstract void onCallDataUsageChanged(long dataUsage);
 
             /**
              * Handles a change in camera capabilities.
